Remove react/dom-server from client bundle
