+++ packages/bodiless-navigation/src/Breadcrumbs/BreadcrumbStoreProvider.tsx
@@ -13,8 +13,6 @@
  */
 
 import React, { ComponentType } from 'react';
-import { useNode, NodeProvider } from '@bodiless/data';
-import { HOC } from '@bodiless/fclasses';
 import type { BreadcrumbStoreType } from './BreadcrumbStore';
 
 const BreadcrumbsStoreContext = React.createContext<BreadcrumbStoreType | undefined>(undefined);
@@ -32,39 +30,7 @@ const BreadcrumbStoreProvider: ComponentType<any> = ({ children, store }: any) =
  */
 const useBreadcrumbStore = () => React.useContext(BreadcrumbsStoreContext);
 
-/**
- * @private
- *
- * Wrap a breadcrumb source which is rendered only on the server to populate
- * the store before any components which need breadcrumb store data are
- * rendered.
- *
- * @param Component
- */
-const asHiddenBreadcrumbSource: HOC = Component => {
-  if (typeof window === 'undefined') {
-    const AsHiddenBreadcrumbSource = (props: any) => {
-      const store = useBreadcrumbStore();
-      const { node } = useNode();
-      // eslint-disable-next-line global-require
-      const reactDomServer = require('react-dom/server');
-
-      reactDomServer.renderToPipeableStream(
-        <NodeProvider node={node}>
-          <BreadcrumbStoreProvider store={store}>
-            <Component {...props} />
-          </BreadcrumbStoreProvider>
-        </NodeProvider>,
-      );
-      return null;
-    };
-    return AsHiddenBreadcrumbSource;
-  }
-  return () => null;
-};
-
 export {
   useBreadcrumbStore,
   BreadcrumbStoreProvider,
-  asHiddenBreadcrumbSource,
 };
