+++ packages/bodiless-navigation/src/Breadcrumbs/asHiddenBreadcrumbSource.static.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { HOC } from '@bodiless/fclasses';
+
+const asHiddenBreadcrumbSource: HOC = Component => (props: any) => <Component {...props} />;
+
+export default asHiddenBreadcrumbSource;
